fix(ci): resolve Windows linker ODR violation + consolidate TriBtn widget
Two NereusSDR::TriBtn classes existed (header-only in TriBtn.h, declared
in RxApplet.h with definitions in RxApplet.cpp) causing a multiple-
definition link error on MinGW. Consolidate into a single .h/.cpp pair,
remove anonymous-namespace copies from FmApplet/PhoneCwApplet, and
switch Windows CI from MinGW Makefiles to Ninja (matching release.yml).
